requirements.txt
awscli==1.29.62
boto3==1.28.62
botocore==1.31.62
colorama==0.4.4
docutils==0.16
jmespath==1.0.1
mysql-connector-python==8.1.0
numpy==1.26.0
protobuf==4.21.12
pyasn1==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==5.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.17
